regex_transformer_django/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-dotenv==1.1.1
openai==1.107.1
pandas==2.3.2
openpyxl==3.1.5
numpy==2.3.3
pytz==2025.2
tzdata==2025.2
